Machine Learning/Projects/Content Engine/requirements.txt
streamlit
langchain
pinecone-client
sentence-transformers
torch
PyMuPDF
